Circular Pendulum Mastery Lab
The purpose of this lab was to create a uniform circular motion pendulum using a string and pendulum bob, and altering certain components of the system to see what would alter the period (time it takes to make one revolution). The components we could've altered were the bob mass, string length, and the path's circumference.