Springs Mastery Lab Progress
Our second and final mastery lab for Physics 3A. Which involves finding the spring constant coefficient of a spring, in simple harmonic motion and displacement alone, under the force of gravity and different masses. The main focus of this lab is to see if there is any distinguishability between the spring constant of a spring under displacement (dead hang) and simple harmonic motion. This is my initial progress of the lab which just entails an overview of the lab's objectives and procedure.