Rough Draft of Spring Mastery Lab
Added final touches to the analysis section of the lab, which required displaying a result for the spring constant taken from the data points when the spring was in simple harmonic motion. Also added the final t prime argument, which stated that the measurements of our spring constants (shm/static) were indistinguishable.